requirements.txt
absl-py==2.3.1
anyio==4.12.0
certifi==2025.11.12
click==8.3.1
colorama==0.4.6
debugpy==1.8.17
exceptiongroup==1.3.1
filelock==3.19.1
fsspec==2025.9.0
ftfy==6.3.1
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.7
idna==3.11
Jinja2==3.1.6
Markdown==3.10
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
protobuf==6.33.1
PyWavelets==1.8.0
PyYAML==6.0.3
regex==2025.11.3
safetensors==0.7.0
sentencepiece==0.2.1
shellingham==1.5.4
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
timm==1.0.22
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
triton==3.3.1
typer-slim==0.20.0
typing_extensions==4.15.0
wcwidth==0.2.14
Werkzeug==3.1.4
